chore: refactor move Operator class into separate file
